Move some namespaces on Storage
